// src/events/types.ts
export enum EventType {
  CONNECT = 'connect',
  HANDSHAKE = 'handshake',
  DISCONNECT = 'diconnect',
  CALL_USER = 'callUser',
  ACCEPT_CALL = 'acceptCall',
  CALL_ACCEPTED = 'callAccepted',
  JOIN_STATUS = 'status',
  HEY = 'hey',
  PLAYER_JOINS_GAME = 'playerJoinsGame',
  PLAYER_JOINED_ROOM = 'playerJoinedRoom',
  CREATE_NEW_GAME = 'createNewGame',
  START_GAME = 'startGame',
  GAME_END = 'endGame',
  NEW_MOVE = 'newMove',
  OPPONENT_MOVE = 'opponentMove',
  REQUEST_USERNAME = 'requestUserName',
  RECEIVED_USERNAME = 'receivedUserName',
  GIVE_USERNAME = 'giveUserName',
  GET_OPPONENT_USERNAME = 'getOpponentUserName',
}